requirements.txt
# Digital Twin System - Python Dependencies
# Cài đặt: pip install -r requirements.txt

# HTTP Client for REST API
requests>=2.31.0

# System and Process monitoring
psutil>=5.9.0

# MQTT Client
paho-mqtt>=1.6.1

# Optional: MongoDB client (nếu muốn kết nối trực tiếp)
# pymongo>=4.6.0
